requirements.txt
Flask==2.2.3
Werkzeug==2.2.3
supabase>=1.0.0
python-dotenv==1.0.0
gunicorn==20.1.0